Add files via upload
PDGFRα+ cells are a specialised population of cells found lying in the interstitial space of visceral organs of the GI Tract and the urinary bladder. They form a network of cells with smooth muscle cells and interstitial cells of Cajal, and this heterogeneous syncytium is known as the SIP syncytium in literature. Cells in these networks interact with one another electrically via gap junctions. As well, the ICC and PDGFRα+ cells have the mechanisms to integrate external stimuli which are both neural and mechanical in origin. More specifically, PDGFRα+ cells are known to mediate inhibitory purinergic neurotransmission, which is a unique feature of this cell type. Here, as part of this thesis, we aim to explore two pathways, (i) coupling of P2Y receptors with SK3 channels and (ii) and coupling stretch sensitive Piezo1 and TRPV4 channels with SK3 channels, towards generation of inhibitory signals when presented with purinergic and mechanical stimuli. This report details a detailed literature survey for the two aforementioned signaling pathways, and also included computational studies of the first pathway using the framework described by Yeoh et al., 2016. The model was implemented on MATLAB and a detailed analysis of this model was performed, its advantages and its limitations were ascertained, and as well of the mechanisms present on PDGFRα+ cells not captured by this model was studied from literature. Finally, a conceptual framework including new mechanisms which include the TRPV4-SK3 coupling. Piezo1-SK3 coupling and other ionic conductances has been proposed as part of future work.